requirements.txt
azure-cognitiveservices-vision-computervision~=0.9.0
Pillow~=9.2.0
python-dotenv~=0.21.0
msrest~=0.7.1
